Rename lib/io.py module which was giving problems
