+++ ckan/lib/munge.py
@@ -11,7 +11,7 @@ import re
 from six import text_type
 
 from ckan import model
-from ckan.lib.io import decode_path
+from ckan.lib.io_ import decode_path
 
 # Maximum length of a filename's extension (including the '.')
 MAX_FILENAME_EXTENSION_LENGTH = 21
